foresee URL-Option in file adding

+++ src/webapp/streamlit_helper.py
@@ -165,19 +165,23 @@ def _visualize_data(container, mainID, agrovoc, projects=[]):
 
 def _file_upload():
     with st.expander("Add Files to project", expanded=False):
-        uploaded_files = st.file_uploader("Select files to add to project",
+        fileadd = st.radio("Add by", options = ["URL","Upload"], horizontal = True)
+        if fileadd == "Upload":
+            uploaded_files = st.file_uploader("Select files to add to project",
                                          key="uploader"+str(st.session_state["file_uploader_key"]),
                                          accept_multiple_files=True,
                                          )
-        if not uploaded_files == []:
-            for uploaded_file in uploaded_files:
-                st.write(uploaded_file.name)
-            if st.button("Add to project"):
+            if not uploaded_files == []:
                 for uploaded_file in uploaded_files:
-                    with open(uploaded_file.name,"wb") as f:
-                        f.write(uploaded_file.getbuffer())
-                    st.session_state.localproject.uploadFilesFromSession(uploaded_file.name)
-                    st.write("added " + uploaded_file.name)
-                    os.remove(uploaded_file.name)
-                st.session_state["file_uploader_key"] += 1
-                st.rerun()
+                    st.write(uploaded_file.name)
+                if st.button("Add to project"):
+                    for uploaded_file in uploaded_files:
+                        with open(uploaded_file.name,"wb") as f:
+                            f.write(uploaded_file.getbuffer())
+                        st.session_state.localproject.uploadFilesFromSession(uploaded_file.name)
+                        st.write("added " + uploaded_file.name)
+                        os.remove(uploaded_file.name)
+                    st.session_state["file_uploader_key"] += 1
+                    st.rerun()
+        if fileadd == "URL":
+            st.write("To be implemented")
